Created server connectivity example to build other funcs
Got rid of dead cryptography hashmaps in main and created a small server connectivity example to list tables current in a DynamoDB structure on AWS. Took the exmaple and made a function based on it as well as small test for printed output.

TODO: Need to correct test because Rust guidelines specify that print statements are for individual use, and that unit tests should, instead, use assert_eq! or something similar.