Corrected code breaking error in lib.rs. Edited the key schema/attributes for create_table and put/delete item functions so that they were of the correct form. Removed all tables that were on AWS System to make sure all tested tables followed correct schema.
There also existed a code-breaking error for create_table. An additional key schema field was left behind as an artifiact of previous code version so a new table in a new region could not be created.
Edited this out; now adding Tests to Sprint Two Test Plan and am editing the main function to accept a  function.